Perform deep clean up
