+++ oai_agents/common/arguments.py
@@ -44,6 +44,14 @@ def get_arguments(additional_args=[]):
     parser.add_argument('--teammates-len',  type=int)
     parser.add_argument('--overcooked-verbose', type=bool, default=False, help="Disables the overcooked game logs")
     
+    parser.add_argument('--pop-total-training-timesteps', type=int)
+    parser.add_argument('--fcp-total-training-timesteps', type=int)
+    parser.add_argument('--fcp-w-sp-total-training-timesteps', type=int)
+    
+    parser.add_argument('--fcp-train-types', nargs='+', type=str)
+    parser.add_argument('--fcp-eval-types', type=dict)
+
+
     for parser_arg, parser_kwargs in additional_args:
         parser.add_argument(parser_arg, **parser_kwargs)
