+++ scripts/train_agents.py
@@ -1,145 +1,51 @@
-import multiprocessing as mp
-mp.set_start_method('spawn', force=True) # should be called before any other module imports
-
-from oai_agents.agents.rl import RLAgentTrainer
 from oai_agents.common.arguments import get_arguments
-from oai_agents.common.tags import TeamType, EvalMembersToBeLoaded
-from scripts.utils import get_fcp_population, update_tms_clction_with_selfplay_types, load_agents, print_teammates_collection
+from oai_agents.common.tags import TeamType
+from utils import get_selfplay_agent, get_fcp_agent_w_tms_clction, get_eval_types_to_load, get_fcp_trained_w_selfplay_types
 
 
-def get_selfplay_agent(args, total_training_timesteps, tag=None, force_training=False):
-    name = 'sp'
-    agents = load_agents(args, name=name, tag=tag, force_training=force_training)
-    if agents:
-        return agents[0]
-
-    selfplay_trainer = RLAgentTrainer(
-        name=name,
-        args=args,
-        agent=None,
-        teammates_collection={},
-        epoch_timesteps=args.epoch_timesteps,
-        n_envs=args.n_envs,
-        seed=678,
-    )
-
-    selfplay_trainer.train_agents(total_train_timesteps=total_training_timesteps)
-    return selfplay_trainer.get_agents()
+def SP(args, pop_force_training):
+    get_selfplay_agent(args=args,
+                    total_training_timesteps=args.pop_total_training_timesteps,
+                    force_training=pop_force_training)
 
 
-def get_fcp_agent_w_tms_clction(args, 
-                                pop_total_training_timesteps,
-                                fcp_total_training_timesteps,
-                                fcp_train_types,
-                                fcp_eval_types,
-                                pop_force_training,
-                                fcp_force_training,
-                                num_self_play_agents_to_train=2,
-                                tag=None,
-                                parallel=True):
-
-    teammates_collection = get_fcp_population(args,
-                                              ck_rate=pop_total_training_timesteps // 5,
-                                              train_types = fcp_train_types,
-                                              eval_types_to_generate = fcp_eval_types['generate'],
-                                              eval_types_to_load_from_file = fcp_eval_types['load'],
-                                              num_self_play_agents_to_train=num_self_play_agents_to_train,
-                                              total_training_timesteps = pop_total_training_timesteps,
-                                              force_training=pop_force_training,
-                                              parallel=parallel)
-    name = 'fcp' 
-    agents = load_agents(args, name=name, tag=tag, force_training=fcp_force_training)
-    if agents:
-        return agents[0], teammates_collection
-
-    fcp_trainer = RLAgentTrainer(
-        name=name,
-        args=args,
-        agent=None,
-        teammates_collection=teammates_collection,
-        epoch_timesteps=args.epoch_timesteps,
-        n_envs=args.n_envs,
-        seed=2602,
-    )
-
-    fcp_trainer.train_agents(total_train_timesteps=fcp_total_training_timesteps)
-    return fcp_trainer.get_agents()[0], teammates_collection
+def FCP(args, pop_force_training, fcp_force_training, parallel):
+    args.fcp_train_types = [TeamType.HIGH_FIRST]
+    args.fcp_eval_types = {'generate' : [TeamType.HIGH_FIRST, TeamType.MEDIUM_FIRST, TeamType.LOW_FIRST],
+                            'load': get_eval_types_to_load()}
+    _, _ = get_fcp_agent_w_tms_clction(args,
+                                        pop_total_training_timesteps=args.pop_total_training_timesteps,
+                                        fcp_total_training_timesteps=args.fcp_total_training_timesteps,
+                                        fcp_train_types=args.fcp_train_types,
+                                        fcp_eval_types=args.fcp_eval_types,
+                                        pop_force_training=pop_force_training,
+                                        fcp_force_training=fcp_force_training,
+                                        num_self_play_agents_to_train=args.num_sp_agents_to_train,
+                                        parallel=parallel)
 
 
-def get_fcp_agent_trained_with_selfplay_types(args,
-                                              fcp_agent,
-                                              tms_clction,
-                                              total_training_timesteps, 
-                                              train_types,
-                                              eval_types,
-                                              force_training=False,
-                                              tag=None):
-    name = 'fcp_w_selfplay_types'
-    agents = load_agents(args, name=name, tag=tag, force_training=force_training)
-    if agents:
-        return agents[0]
-
-    teammates_collection = update_tms_clction_with_selfplay_types(teammates_collection=tms_clction,
-                                                                  agent=fcp_agent,
-                                                                  args=args)
-    fcp_trainer = RLAgentTrainer(
-        name=name,
-        args=args,
-        agent=fcp_agent,
-        teammates_collection=teammates_collection,
-        epoch_timesteps=args.epoch_timesteps,
-        n_envs=args.n_envs,
-        seed=2602,
-    )
-
-    fcp_trainer.train_agents(total_train_timesteps=total_training_timesteps)
-    return fcp_trainer.get_agents()[0]
+def FCP_w_SP_TYPES(args, pop_force_training, fcp_force_training, fcp_w_sp_force_training, parallel):
+    args.fcp_train_types = [TeamType.HIGH_FIRST, TeamType.MEDIUM_FIRST, TeamType.LOW_FIRST]
+    args.fcp_eval_types = {'generate' : [TeamType.HIGH_FIRST, TeamType.MEDIUM_FIRST, TeamType.MIDDLE_FIRST],
+                           'load': []}
+    args.fcp_w_sp_train_types = [TeamType.SELF_PLAY_LOW, TeamType.SELF_PLAY_MEDIUM, TeamType.SELF_PLAY_HIGH]
+    args.fcp_w_sp_eval_types = {'generate': [TeamType.SELF_PLAY_LOW, TeamType.SELF_PLAY_MEDIUM, TeamType.SELF_PLAY_HIGH],
+                                'load': []}
+    get_fcp_trained_w_selfplay_types(args=args,
+                                    pop_total_training_timesteps=args.pop_total_training_timesteps,
+                                    fcp_total_training_timesteps=args.fcp_total_training_timesteps,
+                                    fcp_w_sp_total_training_timesteps=args.fcp_w_sp_total_training_timesteps,
+                                    fcp_train_types=args.fcp_train_types,
+                                    fcp_eval_types=args.fcp_eval_types,
+                                    fcp_w_sp_train_types=args.fcp_w_sp_train_types,
+                                    fcp_w_sp_eval_types=args.fcp_w_sp_eval_types,
+                                    pop_force_training=pop_force_training,
+                                    fcp_force_training=fcp_force_training,
+                                    fcp_w_sp_force_training=fcp_w_sp_force_training,
+                                    parallel=parallel)
 
 
-def train_fcp_and_fcp_w_selfplay(args,
-                                 pop_total_training_timesteps,
-                                 fcp_total_training_timesteps,
-                                 fcp_w_sp_total_training_timesteps,
-                                 pop_force_training,
-                                 fcp_force_training,
-                                 fcp_w_sp_force_training,
-                                 num_self_play_agents_to_train=2,
-                                 parallel=True):
-
-    fcp_train_types = [TeamType.HIGH_FIRST]
-    fcp_eval_types = {
-                        'generate' : [TeamType.HIGH_FIRST, TeamType.MEDIUM_FIRST, TeamType.MIDDLE_FIRST, 
-                                    TeamType.LOW_FIRST, TeamType.RANDOM, TeamType.HIGH_MEDIUM,
-                                    TeamType.HIGH_LOW, TeamType.MEDIUM_LOW, TeamType.HIGH_LOW_RANDOM],
-                        'load': [] # [(TeamType.HIGH_FIRST, name, tag)]
-                    }
-    fcp_agent, teammates_collection = get_fcp_agent_w_tms_clction(args, 
-                                                                  pop_total_training_timesteps=pop_total_training_timesteps,
-                                                                  fcp_total_training_timesteps=fcp_total_training_timesteps,
-                                                                  fcp_train_types=fcp_train_types,
-                                                                  fcp_eval_types=fcp_eval_types,
-                                                                  pop_force_training=pop_force_training,
-                                                                  fcp_force_training=fcp_force_training,
-                                                                  num_self_play_agents_to_train=num_self_play_agents_to_train,
-                                                                  parallel=parallel)
-
-
-    fcp_w_sp_train_types = [TeamType.SELF_PLAY_LOW, TeamType.SELF_PLAY_MEDIUM, TeamType.SELF_PLAY_HIGH]
-    fcp_w_sp_eval_types = {
-                            'generate': [TeamType.SELF_PLAY_LOW, TeamType.SELF_PLAY_MEDIUM, TeamType.SELF_PLAY_HIGH],
-                            'load': []
-                        }
-    get_fcp_agent_trained_with_selfplay_types(args,
-                                              fcp_agent=fcp_agent, 
-                                              tms_clction=teammates_collection,
-                                              total_training_timesteps=fcp_w_sp_total_training_timesteps,
-                                              train_types=fcp_w_sp_train_types,
-                                              eval_types=fcp_w_sp_eval_types,
-                                              force_training=fcp_w_sp_force_training,
-                                              )
-
-
-def get_input(args, quick_test=False):
+def set_input(args, quick_test=False):
     args.layout_names = ['3_chefs_small_kitchen']
     args.teammates_len = 2
     args.num_players = args.teammates_len + 1  # 3 players = 1 agent + 2 teammates
@@ -147,87 +53,50 @@ def get_input(args, quick_test=False):
     if not quick_test: 
         args.n_envs = 50
         args.epoch_timesteps = 1e5
-        pop_total_training_timesteps = 5e6
-        fcp_total_training_timesteps = 5e6
-        fcp_w_sp_total_training_timesteps = 2 * 5e6
+        args.pop_total_training_timesteps = 5e6
+        args.fcp_total_training_timesteps = 5e6
+        args.fcp_w_sp_total_training_timesteps = 2 * 5e6
+        args.num_sp_agents_to_train = 5
 
     else: # Used for doing quick tests
         args.sb_verbose = 1
         args.wandb_mode = 'disabled'
         args.n_envs = 2
         args.epoch_timesteps = 2
-        pop_total_training_timesteps = 3500
-        fcp_total_training_timesteps = 3500
-        fcp_w_sp_total_training_timesteps = 3500 * 2
+        args.pop_total_training_timesteps = 3500
+        args.fcp_total_training_timesteps = 3500
+        args.fcp_w_sp_total_training_timesteps = 3500 * 2
+        args.num_sp_agents_to_train = 2
     
-    return pop_total_training_timesteps, fcp_total_training_timesteps, fcp_w_sp_total_training_timesteps
-
-
-
-def get_eval_types_to_load():
-    '''
-    If load_from_pop_structure is False, it means that we are reading independent agents from files.
-    '''
-    t1 = EvalMembersToBeLoaded(
-        load_from_pop_structure = False,
-        names = ['eval/2_chefs/fcp_hd256_seed26', 'eval/2_chefs/fcp_hd256_seed39'],
-        team_type = TeamType.HIGH_FIRST,
-        tags = ['best', 'best'],
-        layout_name = '3_chefs_small_kitchen',
-    )
-
-    '''
-    Pop structure holds the population of agent used for FCP training. 
-    '''
-    t2 = EvalMembersToBeLoaded(
-        load_from_pop_structure = True,
-        names = ['eval/2_chefs/fcp_pop_3_chefs_small_kitchen'],
-        team_type = TeamType.HIGH_FIRST,
-        tags = ['aamas25'],
-        layout_name = '3_chefs_small_kitchen',
-    )
-    return [t1, t2]
-
 
 if __name__ == '__main__':
     args = get_arguments()
     quick_test = True
-    parallel = True
+    parallel = False
+    
     pop_force_training = False
-    fcp_force_training = True
-    fcp_w_sp_force_training = False
+    fcp_force_training = False
+    fcp_w_sp_force_training = True
     
-    pop_total_training_timesteps, fcp_total_training_timesteps, fcp_w_sp_total_training_timesteps = get_input(args=args,
-                                                                                                              quick_test=quick_test)
+    set_input(args=args, quick_test=quick_test)
 
-    # train_fcp_and_fcp_w_selfplay(args=args,
-    #                              pop_total_training_timesteps=pop_total_training_timesteps,
-    #                              fcp_total_training_timesteps=fcp_total_training_timesteps,
-    #                              fcp_w_sp_total_training_timesteps=fcp_w_sp_total_training_timesteps,
-    #                              pop_force_training=pop_force_training,
-    #                              fcp_force_training=fcp_force_training,
-    #                              fcp_w_sp_force_training=fcp_w_sp_force_training,
-    #                              parallel=parallel)
-
-
-    # get_selfplay_agent(args, force_training=True, total_training_timesteps=3500)
-    # args.train_types = [TeamType.HIGH_FIRST, TeamType.MEDIUM_FIRST, TeamType.LOW_FIRST]
     
-    fcp_train_types = [TeamType.HIGH_FIRST]
-    fcp_eval_types = {
-                        'generate' : [], # TeamType.HIGH_FIRST, TeamType.MEDIUM_FIRST,TeamType.LOW_FIRST,
-                        'load': get_eval_types_to_load()
-                    }
-    fcp_agent, teammates_collection = get_fcp_agent_w_tms_clction(args, 
-                                                                  pop_total_training_timesteps=pop_total_training_timesteps,
-                                                                  fcp_total_training_timesteps=fcp_total_training_timesteps,
-                                                                  fcp_train_types=fcp_train_types,
-                                                                  fcp_eval_types=fcp_eval_types,
-                                                                  pop_force_training=pop_force_training,
-                                                                  fcp_force_training=fcp_force_training,
-                                                                  num_self_play_agents_to_train=2,
-                                                                  parallel=parallel)
+    SP(args=args,
+       pop_force_training=pop_force_training)
 
-    # args.train_types = [TeamType.HIGH_FIRST, TeamType.MEDIUM_FIRST, TeamType.LOW_FIRST,
-    #                     TeamType.HIGH_MEDIUM, TeamType.HIGH_LOW, TeamType.MEDIUM_LOW]
-    # get_fcp_agent(args, force_training=False, parallel=True)
+    FCP(args=args,
+        pop_force_training=pop_force_training,
+        fcp_force_training=fcp_force_training,
+        parallel=parallel)
+
+    FCP_w_SP_TYPES(args=args,
+                   pop_force_training=pop_force_training,
+                   fcp_force_training=fcp_force_training,
+                   fcp_w_sp_force_training=fcp_w_sp_force_training,
+                   parallel=parallel)
+
+
+
+
+
+
