+++ scripts/utils/__init__.py
@@ -1,0 +1,2 @@
+from .train_helper import get_selfplay_agent, get_fcp_agent_w_tms_clction, get_fcp_trained_w_selfplay_types
+from .eval_helper import get_eval_types_to_load
